AdhocMetric
